+++ Assets/Packages/com.studio23.ss2.bettercursormanager/Editor/CursorDataEditorWindow.cs
@@ -9,18 +9,17 @@ namespace Studio23.SS2.BetterCursor.Editor
     public class CursorDataEditorWindow : EditorWindow
     {
         private string _cursorName;
-        private List<Sprite> _cursorTextures=new List<Sprite>(1);
+        private readonly List<Sprite> _cursorTextures = new(1);
         private Vector2 _hotspot = new(.3f, .8f);
         private Vector2 _pixelSize = new(32, 32);
 
-        private float _cursorTextureUpdateDelay=0.1f;
+        private float _cursorTextureUpdateDelay = 0.1f;
 
         private static readonly string _folderPath = "Assets/Resources/BetterCursor/";
 
         [MenuItem("Studio-23/BetterCursor/Create Cursor", false, 1)]
         public static void ShowWindow()
         {
-            
             GetWindow<CursorDataEditorWindow>("Create Cursor");
         }
 
@@ -36,53 +35,43 @@ namespace Studio23.SS2.BetterCursor.Editor
             EditorGUILayout.Space();
             GUILayout.Label("Cursor Textures", EditorStyles.boldLabel);
             // Display existing sprite fields
-            for (int i = 0; i < _cursorTextures.Count; i++)
+            for (var i = 0; i < _cursorTextures.Count; i++)
             {
                 EditorGUILayout.BeginHorizontal();
-                _cursorTextures[i] = (Sprite)EditorGUILayout.ObjectField($"Texture {i + 1}", _cursorTextures[i], typeof(Sprite), true);
+                _cursorTextures[i] =
+                    (Sprite)EditorGUILayout.ObjectField($"Texture {i + 1}", _cursorTextures[i], typeof(Sprite), true);
                 if (GUILayout.Button("X", GUILayout.Width(20)))
                 {
                     _cursorTextures.RemoveAt(i);
                     break;
                 }
+
                 EditorGUILayout.EndHorizontal();
             }
 
             // Add a default sprite field if there are no sprites
-            if (_cursorTextures.Count == 0)
-            {
-                _cursorTextures.Add(null);
-            }
+            if (_cursorTextures.Count == 0) _cursorTextures.Add(null);
 
             // Button to add more sprite fields
-            if (GUILayout.Button("Add Sprite"))
-            {
-                _cursorTextures.Add(null);
-            }
+            if (GUILayout.Button("Add Sprite")) _cursorTextures.Add(null);
 
             EditorGUILayout.Space();
 
-            if (GUILayout.Button("Create Cursor Data"))
-            {
-                CreateCursorDataAsset();
-            }
+            if (GUILayout.Button("Create Cursor Data")) CreateCursorDataAsset();
         }
-
 
 
         private void CreateCursorDataAsset()
         {
-            
-
             var cursorData = CreateInstance<CursorData>();
             cursorData.CursorName = _cursorName;
             cursorData.CursorTextures = _cursorTextures.ToArray();
-            cursorData.TextureUpdateDelay=_cursorTextureUpdateDelay;
+            cursorData.TextureUpdateDelay = _cursorTextureUpdateDelay;
             cursorData.HotSpot = _hotspot;
             cursorData.PixelSize = _pixelSize;
 
 
-            if (!Directory.Exists(_folderPath)) 
+            if (!Directory.Exists(_folderPath))
                 Directory.CreateDirectory(_folderPath);
             var cursorDataPath = Path.Combine(_folderPath, $"{cursorData.name}_CursorData.asset");